+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Initial release! Expect bugs, please report them at [the issue tracker](https://
 	- It makes a sound when it cycles
 - New textures for the Generators and colored blocks
 - Updated many recipes to use data component predicates
+- Magical Snow Globe
+	- Improved tooltip
+    - Now it only requires visiting 7 of the biomes instead of all
 
 ### Fixed